Modificaciones wall
